+++ test/index.js
@@ -5,7 +5,7 @@ const {promisify} = require('util');
 const {join} = require('path');
 const licenseBadger = require('../src/index.js');
 
-const logging = true;
+const logging = 'verbose';
 const packagePath = join(__dirname, '../');
 
 const readFile = promisify(rf);
@@ -30,6 +30,8 @@ const coverageBadgePath = join(__dirname, '../coverage-badge.svg');
 const coverageBadgeFixturePath = getFixturePath('coverage-badge.svg');
 const esmAndMochaPath = getFixturePath('esm-mocha-and-missing.svg');
 const redPublicDomainPath = getFixturePath('redPublicDomain.svg');
+const allDevelopmentPath = getFixturePath('allDevelopment.svg');
+const productionPath = getFixturePath('production.svg');
 const nonemptyFilteredTypes = getFixturePath('nonemptyFilteredTypes.svg');
 const seeLicenseInPath = getFixturePath('seeLicenseIn.svg');
 const unlicensedPath = getFixturePath('unlicensed.svg');
@@ -37,6 +39,7 @@ const reuseProtectivePath = getFixturePath('reuseProtective.svg');
 const uncategorizedPath = getFixturePath('uncategorized.svg');
 
 describe('Main file', function () {
+  this.timeout(5000);
   it('should throw with a bad output path', async () => {
     let err;
     try {
@@ -60,10 +63,16 @@ describe('Main file', function () {
   });
 
   describe('Main functionality', function () {
-    this.timeout(5000);
     const fixturePaths = [];
-    for (let i = 0; i <= 6; i++) {
+    for (let i = 0; i <= 8; i++) {
       fixturePaths.push(join(__dirname, `fixtures/temp${i}.svg`));
+    }
+    let j = 0;
+    /**
+     * @returns {void}
+     */
+    function getNextFixturePath () {
+      return fixturePaths[j++];
     }
     const unlinker = async () => {
       try {
@@ -75,93 +84,129 @@ describe('Main file', function () {
     before(unlinker);
     after(unlinker);
     it('should work with string text color', async function () {
+      const outputPath = getNextFixturePath();
       await licenseBadger({
         packagePath,
         licenseInfoPath,
-        outputPath: fixturePaths[0],
+        outputPath,
         textColor: 'orange,s{blue}',
         logging
       });
-      const contents = await readFile(fixturePaths[0], 'utf8');
+      const contents = await readFile(outputPath, 'utf8');
       const expected = await readFile(esmAndMochaPath, 'utf8');
       expect(contents).to.equal(expected);
     });
 
     it('should work with license type color', async function () {
+      const outputPath = getNextFixturePath();
       await licenseBadger({
         packagePath,
         licenseInfoPath,
-        outputPath: fixturePaths[1],
+        outputPath,
         licenseTypeColor: ['publicDomain=red,s{white}'],
         logging
       });
-      const contents = await readFile(fixturePaths[1], 'utf8');
+      const contents = await readFile(outputPath, 'utf8');
       const expected = await readFile(redPublicDomainPath, 'utf8');
       expect(contents).to.equal(expected);
     });
 
     it('should work with `filteredTypes`', async function () {
+      const outputPath = getNextFixturePath();
       await licenseBadger({
         packagePath,
         licenseInfoPath,
-        outputPath: fixturePaths[2],
+        outputPath,
         filteredTypes: 'nonempty',
         logging
       });
-      const contents = await readFile(fixturePaths[2], 'utf8');
+      const contents = await readFile(outputPath, 'utf8');
       const expected = await readFile(nonemptyFilteredTypes, 'utf8');
       expect(contents).to.equal(expected);
     });
 
-    it('should work with "see license in"', async function () {
+    it('should work with `allDevelopment`', async function () {
+      const outputPath = getNextFixturePath();
       await licenseBadger({
-        corrections: true,
+        allDevelopment: true,
         packagePath,
-        licenseInfoPath: licenseInfoPathSeeLicenseIn,
-        outputPath: fixturePaths[3],
+        outputPath,
         logging
       });
-      const contents = await readFile(fixturePaths[3], 'utf8');
-      const expected = await readFile(seeLicenseInPath, 'utf8');
+      const contents = await readFile(outputPath, 'utf8');
+      const expected = await readFile(allDevelopmentPath, 'utf8');
       expect(contents).to.equal(expected);
     });
 
-    it('should work with "UNLICENSED"', async function () {
+    it('should work with `production`', async function () {
+      const outputPath = getNextFixturePath();
       await licenseBadger({
-        corrections: true,
+        production: true,
+        licenseInfoPath: '',
         packagePath,
-        licenseInfoPath: licenseInfoPathUnlicensed,
-        outputPath: fixturePaths[4],
+        outputPath,
         logging
       });
-      const contents = await readFile(fixturePaths[4], 'utf8');
-      const expected = await readFile(unlicensedPath, 'utf8');
+      const contents = await readFile(outputPath, 'utf8');
+      const expected = await readFile(productionPath, 'utf8');
       expect(contents).to.equal(expected);
     });
 
-    it('should work with Parity', async function () {
-      await licenseBadger({
-        corrections: true,
-        packagePath,
-        licenseInfoPath: licenseInfoPathReuseProtective,
-        outputPath: fixturePaths[5],
-        logging
+    describe('License categories', function () {
+      it('should work with "see license in"', async function () {
+        const outputPath = getNextFixturePath();
+        await licenseBadger({
+          corrections: true,
+          packagePath,
+          licenseInfoPath: licenseInfoPathSeeLicenseIn,
+          outputPath,
+          logging
+        });
+        const contents = await readFile(outputPath, 'utf8');
+        const expected = await readFile(seeLicenseInPath, 'utf8');
+        expect(contents).to.equal(expected);
       });
-      const contents = await readFile(fixturePaths[5], 'utf8');
-      const expected = await readFile(reuseProtectivePath, 'utf8');
-      expect(contents).to.equal(expected);
-    });
 
-    it('should work with an "uncategorized" license', async function () {
-      await licenseBadger({
-        corrections: true,
-        packagePath,
-        licenseInfoPath: licenseInfoPathUncategorized,
-        outputPath: fixturePaths[6]
+      it('should work with "UNLICENSED"', async function () {
+        const outputPath = getNextFixturePath();
+        await licenseBadger({
+          corrections: true,
+          packagePath,
+          licenseInfoPath: licenseInfoPathUnlicensed,
+          outputPath,
+          logging
+        });
+        const contents = await readFile(outputPath, 'utf8');
+        const expected = await readFile(unlicensedPath, 'utf8');
+        expect(contents).to.equal(expected);
       });
-      const contents = await readFile(fixturePaths[6], 'utf8');
-      const expected = await readFile(uncategorizedPath, 'utf8');
-      expect(contents).to.equal(expected);
+
+      it('should work with Parity', async function () {
+        const outputPath = getNextFixturePath();
+        await licenseBadger({
+          corrections: true,
+          packagePath,
+          licenseInfoPath: licenseInfoPathReuseProtective,
+          outputPath,
+          logging
+        });
+        const contents = await readFile(outputPath, 'utf8');
+        const expected = await readFile(reuseProtectivePath, 'utf8');
+        expect(contents).to.equal(expected);
+      });
+
+      it('should work with an "uncategorized" license', async function () {
+        const outputPath = getNextFixturePath();
+        await licenseBadger({
+          corrections: true,
+          packagePath,
+          licenseInfoPath: licenseInfoPathUncategorized,
+          outputPath
+        });
+        const contents = await readFile(outputPath, 'utf8');
+        const expected = await readFile(uncategorizedPath, 'utf8');
+        expect(contents).to.equal(expected);
+      });
     });
   });
 });